requirements.txt
Django==5.0.7
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pillow==10.4.0
django-filter==24.2
djangorestframework-simplejwt==5.3.1
forex-python==1.8
drf-yasg==1.21.7
black==24.4.2
requests==2.32.3
stripe==10.5.0
celery==5.4.0
redis==5.0.8
eventlet==0.36.1
django-celery-beat==2.6.0
django-cors-headers==4.4.0
setuptools==74.1.2

